honeypotter_pkg/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "honeypotter"
version = "0.2.0"
description = "Honeypotter: synthetic OOD distractor generator (patterns, primitives, textures) with in-dist/ood color modes."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{name = "Your Name", email = "you@example.com"}]
keywords = ["OOD", "robustness", "distractor", "textures", "patterns", "albumentations", "ImageNet", "THINGS"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
    "tqdm>=4.60",
    "pandas>=1.5"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "build>=0.10",
    "twine>=4.0"
]

[project.scripts]
honeypotter = "honeypotter.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
